# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Amazon",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pedidos": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Get list of pedidos",
                "description": "Returns list of pedidos",
                "operationId": "c17d0d9bf2445c1fb80ba484e84d7b3a",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pedido/": {
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "description": "Add pedido",
                "operationId": "5769193bf28e63d6874077ba6e10b250",
                "requestBody": {
                    "description": "Pass pedido info",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "direccion",
                                    "localidad",
                                    "zip",
                                    "personaReceptora"
                                ],
                                "properties": {
                                    "direccion": {
                                        "type": "string",
                                        "format": "text",
                                        "example": ""
                                    },
                                    "localidad": {
                                        "type": "string",
                                        "format": "text",
                                        "example": ""
                                    },
                                    "zip": {
                                        "type": "string",
                                        "format": "text",
                                        "example": ""
                                    },
                                    "personaReceptora": {
                                        "type": "string",
                                        "format": "text",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong data introduced"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pedido/{id}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Get a pedido",
                "description": "Returns one pedido",
                "operationId": "82cf2823b520a86089bc46cd74c203a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Delete existing pedido",
                "description": "Deletes a record and returns no content",
                "operationId": "4e248ed9821c93b668a665134fbe2bb0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/peticiones/{id}": {
            "put": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Update existing pedido",
                "description": "Returns updated pedido data",
                "operationId": "bb8542fd0a647627ff3edf6ad3cdede8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "type": "string",
                                        "format": "text",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/producto/": {
            "post": {
                "tags": [
                    "Productos"
                ],
                "description": "Add productos",
                "operationId": "fbc588df3d7222c4c0441e48e0bc44e1",
                "requestBody": {
                    "description": "Pass producto info",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "precio",
                                    "marca",
                                    "descripcion",
                                    "nombre"
                                ],
                                "properties": {
                                    "precio": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": ""
                                    },
                                    "marca": {
                                        "type": "string",
                                        "format": "text",
                                        "example": ""
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "format": "text",
                                        "example": ""
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "format": "text",
                                        "example": ""
                                    },
                                    "image": {
                                        "type": "multipart/form-data",
                                        "format": "file",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong data introduced"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}